Se actualiza contador de puntos
Ahora el contador de puntos se actualiza según avanza el tiempo de la partida

+++ core/src/com/mygdx/game/Auto.java
@@ -36,7 +36,7 @@ public class Auto {
 		public Rectangle getArea() {
 			return car;
 		}
-		public void sumarPuntos(int pp) {
+		public void sumarPuntos(long pp) {
 			puntos+=pp;
 		}
